+++ doc/whats-new.rst
@@ -23,8 +23,7 @@ v2024.02.0 (unreleased)
 New Features
 ~~~~~~~~~~~~
 
-- Added a simple `nbytes` representation in DataArrays and Dataset `repr` (opt-in).
-  To enable, use `xr.set_options(display_nbytes=True)`
+- Added a simple `nbytes` representation in DataArrays and Dataset `repr`.
   (:issue:`8690`, :pull:`8702`).
   By `Etienne Schalk <https://github.com/etienneschalk>`_.
 - Allow negative frequency strings (e.g. ``"-1YE"``). These strings are for example used